Add files via upload
Bug Fixes: 
Fixed the glitch that was causing the export .pdf function to not have the correct formatting in the file output when compared to the LaTeX code (this issue repeats with other export functions). 
Fixed the glitch that was causing the export Word document to not have any correct formatting at all. 
Removed the "Export LaTeX" button from the main workspace button due it being a programmer's test for quickly check LaTex code.